+++ src/lib/text_view.rs
@@ -27,7 +27,7 @@ pub fn traverse_token<F1, F2, R>(token: &Token, f: &F1, add: &F2) -> R
           F2: Fn(R, R) -> R
 {
     match *token {
-        Token::Str(ref s) => f(s),
+        Token::Str(ref s) | Token::Code(ref s) | Token::CodeBlock(_, ref s) => f(s),
 
         Token::SoftBreak => f(" "),