+++ ChangeLog.md
@@ -3,6 +3,8 @@ ChangeLog
 
 unreleased
 -------------
+* Moved from `pulldown-cmark` to `comrak` for parsing Markdown. This may have some
+  performances drawbacks but allows for a few more features.
 * New option: 
   * `crowbook.files_mean_chapters` allow to enforce that each files
     means a chapter break or to make sure that it doesn't (by default,